Upload Progress
I've figured out why the files weren't showing up. enctype was not set in the form. there was no way to set it in the html so I had to add a small script to change it. 

Right now the plugin doesn't have enough privileges to create a directory for the file. I'll be looking into that over the weekend.